pyproject.toml
[tool.poetry]
name = "affilaffe"
version = "0.1.0"
description = ""
authors = ["Simon Lundström <github-commits@soy.se>"]

[tool.poetry.dependencies]
python = "^3.7"
python-gssapi = "^0.6.4"
python-ldap = "^3.2"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[tool.poetry.scripts]
affilaffe = 'affilaffe:main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

